refactor phase 1
